move response builder to own file
