+++ actix-http/src/lib.rs
@@ -42,6 +42,7 @@ mod message;
 mod payload;
 mod request;
 mod response;
+mod response_builder;
 mod service;
 mod time_parser;
 
@@ -59,7 +60,8 @@ pub use self::http_message::HttpMessage;
 pub use self::message::{Message, RequestHead, RequestHeadType, ResponseHead};
 pub use self::payload::{Payload, PayloadStream};
 pub use self::request::Request;
-pub use self::response::{Response, ResponseBuilder};
+pub use self::response::Response;
+pub use self::response_builder::ResponseBuilder;
 pub use self::service::HttpService;
 
 pub mod http {
